Move MigrationTest.kt inside refreshVersions-core
