+++ plugins/core/src/test/resources/migration.files/app/feature/build.gradle.kts
