+++ plugins/core/src/test/resources/migration.files/buildSrc/src/main/kotlin/Versions.kt
